+++ docs/project/changelog.rst
@@ -32,6 +32,14 @@ notice.
 
 *In development*
 
+Bug fixes
+.........
+
+* Wrapped errors when reading the opening handshake request or response in
+  :exc:`~exceptions.InvalidMessage` so that :func:`~asyncio.client.connect`
+  raises :exc:`~exceptions.InvalidHandshake` or a subclass when the opening
+  handshake fails.
+
 .. _14.1:
 
 14.1